Add files via upload
Input file should include:
FILE1 = 
FILE2 =
GROUP1 = 
GROUP2 = 
ATOMS_OF_INT = 
ATOM_THRESH = 
DELOC_THRESH = 
*********
FILE1 is the file name including the Lowdin and/or Mulliken population data for structure 1
FILE2 is the file name including the Lowdin and/or Mulliken population data for structure 2
GROUP1 is a list of atom numbers delimited by "," which constitutes delocalization group 1
GROUP2 is a list of atom numbers delimited by "," which constitutes delocalization group 2
ATOMS_OF_INT is a list of atom numbers delimited by "," which indicates which specific atom population sums should be starred.
ATOM_THRESH a value from >= 0.0, <= 2.0 which indicates an individual population value of an atom within an orbital for it to be printed.  
DELOC_THRESH a value from >= 0.0 which indicates an individual orbital delocalization value for it to be printed. 
DELOC_THRESH, ATOMS_OF_INT, and ATOM_THRESH do not affect the overall index score.